+++ cmd/anubis/js/main.mjs
@@ -24,7 +24,7 @@ const imageURL = (mood, cacheBuster) =>
   const status = document.getElementById('status');
   const image = document.getElementById('image');
   const title = document.getElementById('title');
-  const spinner = document.getElementById('spinner');
+  const progress = document.getElementById('progress');
   const anubisVersion = JSON.parse(document.getElementById('anubis_version').textContent);
 
   // const testarea = document.getElementById('testarea');
@@ -36,8 +36,7 @@ const imageURL = (mood, cacheBuster) =>
   //   title.innerHTML = "Oh no!";
   //   status.innerHTML = "Checks failed. Please check your browser's settings and try again.";
   //   image.src = imageURL("sad");
-  //   spinner.innerHTML = "";
-  //   spinner.style.display = "none";
+  //   progress.style.display = "none";
   //   return;
   // }
 
@@ -54,8 +53,7 @@ const imageURL = (mood, cacheBuster) =>
       title.innerHTML = "Oh no!";
       status.innerHTML = `Failed to fetch config: ${err.message}`;
       image.src = imageURL("sad", anubisVersion);
-      spinner.innerHTML = "";
-      spinner.style.display = "none";
+      progress.style.display = "none";
       throw err;
     });
 
@@ -64,8 +62,7 @@ const imageURL = (mood, cacheBuster) =>
     title.innerHTML = "Oh no!";
     status.innerHTML = `Failed to resolve check algorithm. You may want to reload the page.`;
     image.src = imageURL("sad", anubisVersion);
-    spinner.innerHTML = "";
-    spinner.style.display = "none";
+    progress.style.display = "none";
     return;
   }
 
@@ -74,6 +71,8 @@ const imageURL = (mood, cacheBuster) =>
   status.appendChild(rateText);
 
   let lastSpeedUpdate = 0;
+  let showingApology = false;
+  const likelihood = Math.pow(16, -rules.report_as);
 
   const t0 = Date.now();
   const { hash, nonce } = await process(
@@ -87,6 +86,26 @@ const imageURL = (mood, cacheBuster) =>
         lastSpeedUpdate = delta;
         rateText.data = `${(iters / delta).toFixed(3)}kH/s`;
       }
+
+      // the probability of still being on the page is (1 - likelihood) ^ iters.
+      // by definition, half of the time the progress bar only gets to half, so
+      // apply a polynomial ease-out function to move faster in the beginning
+      // and then slow down as things get increasingly unlikely. quadratic felt
+      // the best in testing, but this may need adjustment in the future.
+      const probability = Math.pow(1 - likelihood, iters);
+      const distance = (1 - Math.pow(probability, 2)) * 100;
+      progress["aria-valuenow"] = distance;
+      progress.firstElementChild.style.width = `${distance}%`;
+
+      if (probability < 0.1 && !showingApology) {
+        status.append(
+          document.createElement("br"),
+          document.createTextNode(
+            "Verification is taking longer than expected. Please do not refresh the page.",
+          ),
+        );
+        showingApology = true;
+      }
     },
   );
   const t1 = Date.now();
@@ -95,8 +114,7 @@ const imageURL = (mood, cacheBuster) =>
   title.innerHTML = "Success!";
   status.innerHTML = `Done! Took ${t1 - t0}ms, ${nonce} iterations`;
   image.src = imageURL("happy", anubisVersion);
-  spinner.innerHTML = "";
-  spinner.style.display = "none";
+  progress.style.display = "none";
 
   setTimeout(() => {
     const redir = window.location.href;